Ajout du petit manuel d'utilisation, ajout des directives de préprocesseur pour enlever l'affichage. Pour l'instant le makefile contient le flag -DDEBUG, mais il suffit de l'enlever dans le makefile pour que les calculs ne génèrent plus aucun message, modification de la fonction createSortedList pour qu'elle reçoive un tableau en argument

+++ src/Paire.cpp
@@ -14,7 +14,7 @@ Paire::~Paire()
     //dtor
 }
 //surcharge du constructeur
-Paire::Paire (float valeur, int position) : m_longueur(valeur), m_position(position)
+Paire::Paire (double valeur, int position) : m_longueur(valeur), m_position(position)
 {
 
 }
@@ -27,11 +27,11 @@ void Paire::affiche ()
 int Paire::getPos()
 {return m_position;}
 
-float Paire::getLongueur()
+double Paire::getLongueur()
 {return m_longueur;}
 
  void Paire::setPosition (int pos)
  {m_position = pos;}
 
- void Paire::setLongueur (float l)
+ void Paire::setLongueur (double l)
  {m_longueur=l;}
